Value_inputs which don't accept blocks
This is a stub for support of placing blocks into dropdown inputs which
do not accept blocks. This commit works to unify all dropdown inputs by
making them use the same code - that is, the code of value_inputs, and
to never (directly) embed fields.

Inputs may be marked to not take input blocks by setting "acceptsBlocks"
to false under their definition in initJson (under args0, etc).

This commit changes the "abs of" (math operation of) block to use one
such dropdown, and so, with appropriate modification of the sb2 import
specmap, projects which use the hacked "(reporter) of (number)" work.

The only problem with this is backwards compatibility: while I haven't
tested it, I expect that old Scratch 3.0 projects would not work with
these changes (for example any 3.0 project which uses the "abs of" block
would not load correctly). Since a PR such as this would almost
certainly never be merged before the release of Scratch 3.0 (simply
because of the ST prioritizing other more important issues), I'm not
sure what hope there is for hacked argument support in this form.
A PR such as this could not be merged AFTER the release of 3.0 without
being backwards-compatible, since there will be many projects made as
soon as 3.0 released and it would be important to not break any of them.